pyproject.toml
[project]
name = "yolo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mne>=1.11.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "torchinfo>=1.8.0",
    "tqdm>=4.67.3",
]
